+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.1.1] - 2025-09-10
 - fix #3, reading
+- improve error handling
+- add **int getError()**
+- add CRC8 check for receiving data
+- add **CHT832X_ERROR_CRC**
+- add constants for the COMMANDS in .cpp (readability)
 - update GitHub actions
 - minor edits